First version of LeftPanel
This panel will be used for the Commission, Commissioner, and Piece left
panels. This version includes some changes to the way the testmodel
stores data to match the way the Panel requests data. Next is to build
search and insertion into the LeftPanel and then begin working on the
right panels.