Rewrote proxy model
Fixed the bugs that existed by switching it to being QIdentityProxyModel
based. I can now add a search row onto the bottom of the model, but I
need to figure out a way to make it stay in place and not scroll.
I think I may have found a solution, but it might involve some major
changes to the leftproxymodel, hence the commit.