Working with getting Search Box behavior right
Ideal functionality will let tab and backtab navigate to and through the
search box, and enter will submit.
The form for the search box is now complete, so after this comes
creating a custom delegate to show persistent editor buttons on the
rightmost column to allow the user to delete rows. Then I will test to
make sure everything is throwing signals appropriately.

Then, dare I say, done with the left panel?