Converted SearchTable into generic FixedRowTable
Added several features to it like the ability to toggle buttons on/off
button. Currently working on ability to toggle fixed box position;
mostly works but there is an issue where when the box is placed on the
bottom, the search row that is normally hidden by the box at the top of
the table is now visible. Trying to figure out how to hide it in only a
single view, as hideRow() seems to hide the search box as well.